requirements.txt
dabl==0.2.6.2
einops==0.8.0
fbm==0.3.0
klib==1.1.2
lightgbm==4.4.0
matplotlib==3.7.1
numpy==1.24.1
pandas==2.2.2
pmdarima==2.0.4
ptflops==0.7.3
pynvml==11.5.0
python_dateutil==2.8.2
scikit_learn==1.3.1
scipy==1.14.0
seaborn==0.13.2
shap==0.46.0
statsmodels==0.14.2
sympy==1.11.1
torch==2.0.1+cu118
tqdm==4.65.0
tvm==1.0.0
